requirements.txt

python-dotenv
yelpapi
pandas
numpy
openpyxl